Use pathfinding module to build board
Create a new, blank representation of the board - a simple grid of the dimensions specified in the POST request.
By default, every point on the board is initially walkable.
In-code notes explain that a new board needs to be created to handle every request to '/move' because the pathfinding library modifies the board when it accesses it to find a path.