Determine location of own head
Knowing your own snake's starting point (ie head location) is crucial for determining a path and a move.
In theory it would be possible to store this information between moves and update it each turn based on the move that is calculated, but it would be problematic if the server ever failed to receive the chosen move response in time: the server would choose a (random?) move and all future calculations would be thrown off. Better to use the authoritative data provided.