Create finder from pathfinding library
Pathfinding functionality is accessed through the 'finder' object, which can be set to refer to any of the ten pathfinding algorithms made available by the library.
Initially set to 'BestFirstFinder' because is one of the algorithms that could fulfill our needs. Can be changed at any point in the future when the pros/cons of different algorithms become clearer.