Structure for AI init.
The topAI will handle the request.  It will first hand it off to the preprocessor, which will take care of making inferences given the request.  Then the preprocessor returns the board object.  Every AI will take board object as an argument.  The board object will contain the fields and data of the original request as well.

Specs are done with jasmine.