Creates graph object from an nxn grid of values.
These values may be either 0s and 1s denoting walkable and non-walkable states, or other values implying varying difficulties of terrain i.e. the probability of a snake occupying the space.  This function will be useful in rewriting Dijkstra of A* to find shortest path given a changing weighted graphs.  The graph.js object has methods for iterating over all cycles, paths, sinks, etc. in a graph, which may not be useful given they are iterated through in no particular order, and the search space could be large.

It might be important to check how much time any of these methods are taking.  Use Use the Node.js console.time() and console.timeEnd() for this.