Refactor grid creation to use width & height
We will not necessarily be on a square board at all times. This commit refactors the 'createGrid' function so that it takes two parameters, width & height, instead of one 'size' parameter that was used to create an NxN square grid. The iteration variables used inside the function are also switched to correspond more closely with conventional use of 'x' relating to horizontal width and 'y' relating to vertical height.