Create a grid without snake auras
Some AI functionality will likely need a grid that doesn't have snake auras on it. This commit refactors the Board constructor to also create a
'snakes only' board that should only show snake placement. Required a minor refactor of the 'addSnakes' function to add snakes to the second grid
while adding them to the main grid."