Snake class changed so that snakes can be cloned.  Added move method.
The idea here is that if we want to explore a subset of future possible moves, we have to have a snake object that (a) can move and (b) doesn't modify the references to data on the board object about the current snake.   The way I solved this was by implementing a move method on the Snake class, so that any given snake instance can be given a move, and return a deep clone of that snake that has made the move.

TODO: to use this, we might want to instantiant extra board instances representing future possible boards.